NEW COMMIT

+++ src/test/java/model/Main.java
@@ -4,8 +4,8 @@ public class Main {
     private long temp;
     private int pressure;
     private int humidity;
-    private long temp_min;
-    private long temp_max;
+    private double temp_min;
+    private double temp_max;
 
     public long getTemp() {
         return temp;
@@ -31,19 +31,19 @@ public class Main {
         this.humidity = humidity;
     }
 
-    public long getTemp_min() {
+    public double getTemp_min() {
         return temp_min;
     }
 
-    public void setTemp_min(long temp_min) {
+    public void setTemp_min(double temp_min) {
         this.temp_min = temp_min;
     }
 
-    public long getTemp_max() {
+    public double getTemp_max() {
         return temp_max;
     }
 
-    public void setTemp_max(long temp_max) {
+    public void setTemp_max(double temp_max) {
         this.temp_max = temp_max;
     }
 }
